chore: Update dependencies and modularize codebase for v5.7.0
This release updates all npm dependencies to their latest versions and completes
the comprehensive code refactoring to comply with coding standards.

## Dependency Updates
- Next.js: 16.0.0 → 16.0.3
- Zod: 3.25.76 → 4.1.12 (MAJOR - with migration fixes)
- dexie-react-hooks: 1.1.7 → 4.2.0 (MAJOR)
- @types/node: 20.x → 24.x (MAJOR)
- @vitejs/plugin-react: 4.7.0 → 5.1.1 (MAJOR)
- jsdom: 24.1.3 → 27.2.0 (MAJOR)
- All vitest packages: 4.0.8 → 4.0.9
- All other dependencies updated to latest stable versions
- Removed deprecated @types/testing-library__jest-dom

## Code Modularization
- Split lib/analytics.ts (333 lines) into 5 focused modules
- Split lib/notifications.ts (342 lines) into 5 focused modules
- Split components/task-form.tsx (355 lines) into 3 modules
- Refactored components/sync/sync-button.tsx (342 → 194 lines)
- Split MCP server analytics.ts (419 lines) into 5 modules
- Split MCP server cli.ts (420 lines) into 4 modules
- Split MCP server tools.ts (439 lines) into 10 modules
- Created lib/constants/sync.ts for centralized constants
- Added comprehensive "why" comments to complex algorithms

## Zod v4 Migration
- Updated error.errors → error.issues in oauth-schemas.ts
- Changed errorMap parameter to error parameter
- All validation logic updated for Zod v4 compatibility

## Version Updates
- Main app: 5.6.0 → 5.7.0
- MCP server: 0.5.1 → 0.6.0

All files now comply with coding standards (<300 lines per file, <30 lines per function).
Zero breaking changes - 100% backward compatibility maintained via re-export layers.
Production build verified successful. MCP server tests: 64/64 passing.
