+++ components/task-form/index.tsx
@@ -1,8 +1,5 @@
 "use client";
 
-import { useState } from "react";
-import { z } from "zod";
-import { taskDraftSchema } from "@/lib/schema";
 import type { TaskDraft, RecurrenceType } from "@/lib/types";
 import { TogglePill } from "@/components/toggle-pill";
 import { TaskFormTags } from "@/components/task-form-tags";
@@ -12,6 +9,8 @@ import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
 import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
+import { useTaskForm } from "./use-task-form";
+import { TIME_OPTIONS, isoToDateInput } from "./validation";
 
 interface TaskFormProps {
   taskId?: string; // ID of task being edited (undefined for new tasks)
@@ -22,141 +21,24 @@ interface TaskFormProps {
   submitLabel?: string;
 }
 
-interface FormErrors {
-  title?: string;
-  description?: string;
-  dueDate?: string;
-  tags?: string;
-  subtasks?: string;
-}
-
-const defaultValues: TaskDraft = {
-  title: "",
-  description: "",
-  urgent: true,
-  important: true,
-  dueDate: undefined,
-  recurrence: "none",
-  tags: [],
-  subtasks: [],
-  dependencies: [],
-  notifyBefore: 15, // Default to 15 minutes before
-  notificationEnabled: true
-};
-
-// Generate time options in 15-minute increments with 12-hour AM/PM format
-function generateTimeOptions(): Array<{ value: string; label: string }> {
-  const options: Array<{ value: string; label: string }> = [];
-  for (let hour = 0; hour < 24; hour++) {
-    for (let minute = 0; minute < 60; minute += 15) {
-      const h24 = hour.toString().padStart(2, '0');
-      const m = minute.toString().padStart(2, '0');
-      const value = `${h24}:${m}`;
-
-      // Convert to 12-hour format
-      const period = hour >= 12 ? 'PM' : 'AM';
-      const h12 = hour === 0 ? 12 : hour > 12 ? hour - 12 : hour;
-      const label = `${h12}:${m} ${period}`;
-
-      options.push({ value, label });
-    }
-  }
-  return options;
-}
-
-const TIME_OPTIONS = generateTimeOptions();
-
-function isoToDateInput(isoValue?: string): string {
-  if (!isoValue) {
-    return "";
-  }
-  const date = new Date(isoValue);
-  if (Number.isNaN(date.getTime())) {
-    return "";
-  }
-  return date.toISOString().slice(0, 10);
-}
-
-function isoToTimeInput(isoValue?: string): string {
-  if (!isoValue) {
-    return "";
-  }
-  const date = new Date(isoValue);
-  if (Number.isNaN(date.getTime())) {
-    return "";
-  }
-  // Get local time in HH:MM format
-  const hours = date.getHours().toString().padStart(2, '0');
-  const minutes = date.getMinutes().toString().padStart(2, '0');
-  return `${hours}:${minutes}`;
-}
-
-function dateTimeInputToIso(dateValue?: string, timeValue?: string): string | undefined {
-  if (!dateValue) {
-    return undefined;
-  }
-  // If time is provided, use it; otherwise default to start of day
-  const time = timeValue || "00:00";
-  const iso = new Date(`${dateValue}T${time}:00`).toISOString();
-  return iso;
-}
-
 export function TaskForm({
   taskId,
-  initialValues = defaultValues,
+  initialValues,
   onSubmit,
   onCancel,
   onDelete,
   submitLabel = "Save task"
 }: TaskFormProps) {
-  const [values, setValues] = useState<TaskDraft>({ ...defaultValues, ...initialValues });
-  const [errors, setErrors] = useState<FormErrors>({});
-  const [submitting, setSubmitting] = useState(false);
-  const [selectedTime, setSelectedTime] = useState<string>(isoToTimeInput(initialValues.dueDate));
-
-  const updateField = <Key extends keyof TaskDraft>(key: Key, value: TaskDraft[Key]) => {
-    setValues((current) => ({ ...current, [key]: value }));
-  };
-
-  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    setSubmitting(true);
-    setErrors({});
-
-    try {
-      const normalized: TaskDraft = {
-        ...values,
-        dueDate: dateTimeInputToIso(isoToDateInput(values.dueDate), selectedTime)
-      };
-      const validated = taskDraftSchema.parse({
-        ...normalized,
-        description: normalized.description ?? ""
-      });
-      await onSubmit(validated);
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        const fieldErrors: FormErrors = {};
-        for (const issue of error.issues) {
-          if (issue.path[0] === "title") {
-            fieldErrors.title = issue.message;
-          }
-          if (issue.path[0] === "description") {
-            fieldErrors.description = issue.message;
-          }
-          if (issue.path[0] === "dueDate") {
-            fieldErrors.dueDate = "Please choose a valid date";
-          }
-        }
-        setErrors(fieldErrors);
-      }
-      // Note: onSubmit handler in parent component will handle non-validation errors
-      setSubmitting(false);
-      return;
-    }
-
-    setSubmitting(false);
-    onCancel();
-  };
+  const {
+    values,
+    errors,
+    submitting,
+    selectedTime,
+    updateField,
+    updateTime,
+    updateDate,
+    handleSubmit
+  } = useTaskForm({ initialValues, onSubmit, onCancel });
 
   return (
     <form className="space-y-4" onSubmit={handleSubmit}>
@@ -230,23 +112,13 @@ export function TaskForm({
             id="due-date"
             type="date"
             value={isoToDateInput(values.dueDate)}
-            onChange={(event) => {
-              const nextIso = dateTimeInputToIso(event.target.value, selectedTime);
-              updateField("dueDate", nextIso);
-            }}
+            onChange={(event) => updateDate(event.target.value)}
           />
           <div className="space-y-1">
             <select
               id="due-time"
               value={selectedTime}
-              onChange={(event) => {
-                setSelectedTime(event.target.value);
-                // Update the dueDate with the new time if a date is already set
-                if (values.dueDate) {
-                  const nextIso = dateTimeInputToIso(isoToDateInput(values.dueDate), event.target.value);
-                  updateField("dueDate", nextIso);
-                }
-              }}
+              onChange={(event) => updateTime(event.target.value)}
               className="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
             >
               <option value="">No specific time</option>
