+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "gsd-taskmanager",
-	"version": "5.6.0",
+	"version": "5.7.0",
 	"private": true,
 	"scripts": {
 		"dev": "next dev",
@@ -32,41 +32,40 @@
 		"clsx": "^2.1.1",
 		"date-fns": "^4.1.0",
 		"dexie": "^4.2.1",
-		"dexie-react-hooks": "^1.1.7",
-		"lucide-react": "^0.546.0",
+		"dexie-react-hooks": "^4.2.0",
+		"lucide-react": "^0.553.0",
 		"nanoid": "^5.1.6",
-		"next": "16.0.0",
+		"next": "16.0.3",
 		"next-themes": "^0.4.6",
 		"react": "^19.2.0",
 		"react-dom": "^19.2.0",
-		"recharts": "^3.3.0",
+		"recharts": "^3.4.1",
 		"sonner": "^2.0.7",
-		"zod": "^3.25.76"
+		"zod": "^4.1.12"
 	},
 	"devDependencies": {
 		"@testing-library/dom": "^10.4.1",
 		"@testing-library/jest-dom": "^6.9.1",
 		"@testing-library/react": "^16.3.0",
 		"@testing-library/user-event": "^14.6.1",
-		"@types/node": "^20.19.22",
-		"@types/react": "^19.2.2",
-		"@types/react-dom": "^19.2.2",
-		"@types/testing-library__jest-dom": "^6.0.0",
-		"@vitejs/plugin-react": "^4.7.0",
-		"@vitest/coverage-v8": "^4.0.3",
-		"autoprefixer": "^10.4.21",
+		"@types/node": "^24.10.1",
+		"@types/react": "^19.2.5",
+		"@types/react-dom": "^19.2.3",
+		"@vitejs/plugin-react": "^5.1.1",
+		"@vitest/coverage-v8": "^4.0.9",
+		"autoprefixer": "^10.4.22",
 		"babel-plugin-react-compiler": "^1.0.0",
 		"eslint": "^9.39.1",
-		"eslint-config-next": "16.0.0",
+		"eslint-config-next": "16.0.3",
 		"fake-indexeddb": "^6.2.4",
-		"jsdom": "^24.1.3",
+		"jsdom": "^27.2.0",
 		"postcss": "^8.5.6",
 		"rimraf": "^6.0.1",
 		"tailwindcss": "^3.4.18",
 		"typescript": "^5.9.3",
 		"vite": "^7.1.12",
-		"vitest": "^4.0.3",
-		"wrangler": "^4.43.0"
+		"vitest": "^4.0.9",
+		"wrangler": "^4.47.0"
 	},
 	"pnpm": {
 		"overrides": {
