+++ packages/mcp-server/src/constants.ts
@@ -1,0 +1,5 @@
+/**
+ * Maximum number of tasks to fetch per pull request
+ * Must match Worker schema validation limit
+ */
+export const MAX_TASKS_PER_PULL = 100;
